refactor(ui): migrate ui primitives to primitives directory
