+++ src/components/marketing/OneApp.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image';
 
-import { Carousel, CarouselContent, CarouselItem } from '@/components/ui/Carousel';
 import { PyroButton } from '@/components/ui/PyroButton';
 import PyroPill from '@/components/ui/PyroPill';
+import { Carousel, CarouselContent, CarouselItem } from '@/components/ui/primitives/Carousel';
 
 const OneApp = () => {
     return (
